Pull out anchor links into their own component
